Refactured information buffering in scan-registration.
Before this refactoring, the scan-registration component used four large static arrays for information buffering. Some information like point curvature, etc. were processed for the whole input cloud and then buffered for later use. As a result, the static arrays must be larger than the maximum number of valid points received within an input cloud, otherwise the component will crash.
On the other hand, usage of the information buffers during later calculations is rather local. The usage of point curvatures, labels and sorting indices is limited to the processing of regions. The usage of the neighbor picked flags is naturally limited to a scanning ring. This refactoring respects the corresponding usage scopes of the buffers, while making them dynamic vectors. This way memory usage is automatically adapted to the ring and ring-region point sizes, making the algorithm effectively independent of the Velodyne lidar in use. As a side effect, overall memory usage should also be reduced.

* Raplaced static information buffer arrays with dynamic vectors.
* Moved calculation of point curvature to processing of region.
* Moved rejection of unreliable points to processing of scan ring.
* Changed the way of determining the (valid) full resolution input cloud size
* Fixed some potential index out of bounds issues by using the CURVATURE_REGION variable to mark regions around unreliable points.