Determine number of scan rings while parsing input cloud.
Up to now, the number of scan rings was hard coded to allow static initialization of buffers. With this commit, all corresponding buffers (ringStartIndices, ringEndIndices and laserCloudScans) are automatically resized if a point is processed which ring index would exceed the buffers.

If everything works as expeced, you should be able now to run the scan-registration component with any Velodyne lidar, without the need to specify the model explicitly. Of course, certain parts like SCAN_PERIOD, etc. may still be sensor specific, but the main algorithm should work fine.